# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AssetCurrency = typeof AssetCurrency[keyof typeof AssetCurrency];


export const AssetCurrency = {
  USD: 'USD',
  Robux: 'Robux',
} as const;

export type AssetCategory = typeof AssetCategory[keyof typeof AssetCategory];


export const AssetCategory = {
  UI: 'UI',
  Scripts: 'Scripts',
  '3D_Models': '3D Models',
  Maps: 'Maps',
} as const;

export interface Asset {
  id: number;
  title: string;
  description: string;
  price: number;
  currency: AssetCurrency;
  category: AssetCategory;
  image_urls: string[];
  discord_link: string;
  created_at: string;
  updated_at: string;
}

export type AssetInputCurrency = typeof AssetInputCurrency[keyof typeof AssetInputCurrency];


export const AssetInputCurrency = {
  USD: 'USD',
  Robux: 'Robux',
} as const;

export type AssetInputCategory = typeof AssetInputCategory[keyof typeof AssetInputCategory];


export const AssetInputCategory = {
  UI: 'UI',
  Scripts: 'Scripts',
  '3D_Models': '3D Models',
  Maps: 'Maps',
} as const;

export interface AssetInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  price: number;
  currency: AssetInputCurrency;
  category: AssetInputCategory;
  /** @minItems 1 */
  image_urls: string[];
  discord_link: string;
}

export type AssetUpdateCurrency = typeof AssetUpdateCurrency[keyof typeof AssetUpdateCurrency];


export const AssetUpdateCurrency = {
  USD: 'USD',
  Robux: 'Robux',
} as const;

export type AssetUpdateCategory = typeof AssetUpdateCategory[keyof typeof AssetUpdateCategory];


export const AssetUpdateCategory = {
  UI: 'UI',
  Scripts: 'Scripts',
  '3D_Models': '3D Models',
  Maps: 'Maps',
} as const;

export interface AssetUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  description?: string;
  /** @minimum 0 */
  price?: number;
  currency?: AssetUpdateCurrency;
  category?: AssetUpdateCategory;
  /** @minItems 1 */
  image_urls?: string[];
  discord_link?: string;
}

export interface MarketplaceSummary {
  asset_count: number;
  category_count: number;
  /** @nullable */
  newest_asset: string | null;
}

export interface AdminStatus {
  configured: boolean;
  authenticated: boolean;
}

export interface AdminPasswordInput {
  /** @minLength 8 */
  password: string;
}

export interface ChangePasswordInput {
  /** @minLength 1 */
  current_password: string;
  /** @minLength 8 */
  new_password: string;
}

export interface Error {
  error: string;
}

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type ListAssetsParams = {
search?: string;
category?: string;
sort?: ListAssetsSort;
};

export type ListAssetsSort = typeof ListAssetsSort[keyof typeof ListAssetsSort];


export const ListAssetsSort = {
  newest: 'newest',
  oldest: 'oldest',
  price_low: 'price_low',
  price_high: 'price_high',
} as const;