# lib/api-zod/src/generated/types/adminStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStatus {
  configured: boolean;
  authenticated: boolean;
}