# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssetCategory } from './assetCategory';
import type { AssetCurrency } from './assetCurrency';

export interface Asset {
  id: number;
  title: string;
  description: string;
  price: number;
  currency: AssetCurrency;
  category: AssetCategory;
  image_urls: string[];
  discord_link: string;
  created_at: Date;
  updated_at: Date;
}