# lib/api-zod/src/generated/types/assetInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssetInputCategory } from './assetInputCategory';
import type { AssetInputCurrency } from './assetInputCurrency';

export interface AssetInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  price: number;
  currency: AssetInputCurrency;
  category: AssetInputCategory;
  /** @minItems 1 */
  image_urls: string[];
  discord_link: string;
}