# lib/api-zod/src/generated/types/assetUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssetUpdateCategory } from './assetUpdateCategory';
import type { AssetUpdateCurrency } from './assetUpdateCurrency';

export interface AssetUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  description?: string;
  /** @minimum 0 */
  price?: number;
  currency?: AssetUpdateCurrency;
  category?: AssetUpdateCategory;
  /** @minItems 1 */
  image_urls?: string[];
  discord_link?: string;
}