# lib/api-zod/src/generated/types/changePasswordInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChangePasswordInput {
  /** @minLength 1 */
  current_password: string;
  /** @minLength 8 */
  new_password: string;
}