# lib/api-zod/src/generated/types/marketplaceSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplaceSummary {
  asset_count: number;
  category_count: number;
  /** @nullable */
  newest_asset: Date | null;
}